# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/broker/projects/{idOrSlug}": {
      "get": {
        "operationId": "BrokerMarketplaceController_detail_v1",
        "parameters": [
          { "name": "idOrSlug", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BrokerProjectResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Broker Marketplace"]
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "ProjectsController_list_v1",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Project, developer or location search",
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "name": "provinceId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "name": "districtId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "name": "developerOrganizationId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": { "example": "6000000.0000", "type": "string" }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": { "example": "12000000.0000", "type": "string" }
          },
          {
            "name": "roomType",
            "required": false,
            "in": "query",
            "schema": { "example": "2+1", "type": "string" }
          },
          {
            "name": "minNetArea",
            "required": false,
            "in": "query",
            "description": "Minimum net area of an available unit",
            "schema": { "example": "80.00", "type": "string" }
          },
          {
            "name": "maxDownPaymentPercent",
            "required": false,
            "in": "query",
            "description": "Maximum down-payment percentage",
            "schema": { "example": "30.00", "type": "string" }
          },
          {
            "name": "maxMonthlyPayment",
            "required": false,
            "in": "query",
            "description": "Maximum monthly payment",
            "schema": { "example": "300000.0000", "type": "string" }
          },
          {
            "name": "deliveryReady",
            "required": false,
            "in": "query",
            "description": "Only ready or only not-ready projects",
            "schema": { "type": "boolean" }
          },
          {
            "name": "deliveryBefore",
            "required": false,
            "in": "query",
            "description": "Delivery on or before this date",
            "schema": { "format": "date", "type": "string" }
          },
          {
            "name": "bounds",
            "required": false,
            "in": "query",
            "description": "west,south,east,north decimal bounds",
            "schema": { "type": "string" }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "NEWEST",
              "type": "string",
              "enum": ["NEWEST", "PRICE_ASC", "PRICE_DESC", "DELIVERY_ASC"]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from pageInfo.nextCursor",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectListResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "tags": ["Projects"]
      }
    },
    "/api/v1/projects/{idOrSlug}": {
      "get": {
        "operationId": "ProjectsController_get_v1",
        "parameters": [
          { "name": "idOrSlug", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectDetailResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "tags": ["Projects"]
      }
    },
...
    "/api/v1/developer/projects/{projectId}": {
      "get": {
        "operationId": "DeveloperProjectsController_detail_v1",
        "parameters": [
          { "name": "projectId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperProjectDetailResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Projects"]
      }
    },
...
  },
  "info": {
    "title": "PLANDA Product API",
    "description": "Shared API for web, iOS, Android, developer, broker and admin clients",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Validation failed" },
          "details": { "type": "array", "items": { "type": "string" } },
          "requestId": { "type": "string", "example": "4f12bc5f-1a5e-4d14-bf14-6c7bb19ea413" }
        },
        "required": ["statusCode", "code", "message", "requestId"]
      },
...
      "Object": { "type": "object", "properties": {} },
...
      "BrokerProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "developerName": { "type": "string" },
          "publicStartingPrice": { "type": "string" },
          "brokerPrice": { "type": "object", "nullable": true },
          "commissionPercent": { "type": "object", "nullable": true },
          "reservationHours": { "type": "object", "nullable": true },
          "salesContact": { "type": "object", "nullable": true },
          "units": { "type": "array", "items": { "type": "object" } },
          "materialCount": { "type": "number" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "slug",
          "name",
          "developerName",
          "publicStartingPrice",
          "units",
          "materialCount",
          "updatedAt"
        ]
      },
...
      "ProjectLocationReferenceDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" }
        },
        "required": ["id", "code", "name", "slug"]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "developerOrganizationId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "developerName": { "type": "string" },
          "developerVerified": { "type": "boolean" },
          "status": { "type": "string", "enum": ["DRAFT", "IN_REVIEW", "PUBLISHED", "ARCHIVED"] },
          "constructionStatus": {
            "type": "string",
            "enum": ["PLANNED", "UNDER_CONSTRUCTION", "READY"]
          },
          "province": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "district": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "latitude": { "type": "string", "example": "40.990868" },
          "longitude": { "type": "string", "example": "29.027707" },
          "startingPrice": { "type": "string", "example": "7500000.0000" },
          "currency": { "type": "string", "example": "TRY" },
          "deliveryDate": { "type": "string", "format": "date", "nullable": true },
          "summary": { "type": "string", "nullable": true },
          "heroImageUrl": { "type": "string", "nullable": true },
          "stockUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "priceUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "version": { "type": "number" }
        },
        "required": [
          "id",
          "developerOrganizationId",
          "name",
          "slug",
          "developerName",
          "developerVerified",
          "status",
          "constructionStatus",
          "province",
          "district",
          "latitude",
          "longitude",
          "startingPrice",
          "currency",
          "deliveryDate",
          "summary",
          "heroImageUrl",
          "stockUpdatedAt",
          "priceUpdatedAt",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
      "PageInfoDto": {
        "type": "object",
        "properties": {
          "nextCursor": { "type": "string", "nullable": true },
          "hasNextPage": { "type": "boolean" }
        },
        "required": ["nextCursor", "hasNextPage"]
      },
      "ProjectListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProjectResponseDto" }
          },
          "pageInfo": { "$ref": "#/components/schemas/PageInfoDto" }
        },
        "required": ["items", "pageInfo"]
      },
      "ProjectUnitTypeDto": {
        "type": "object",
        "properties": {
          "roomType": { "type": "string" },
          "availableCount": { "type": "number" },
          "minNetArea": { "type": "string" },
          "maxNetArea": { "type": "string" },
          "startingPrice": { "type": "string" },
          "currency": { "type": "string" },
          "minFloor": { "type": "number", "nullable": true },
          "maxFloor": { "type": "number", "nullable": true },
          "orientations": { "type": "array", "items": { "type": "string" } },
          "floorPlanImageUrl": { "type": "string", "nullable": true }
        },
        "required": [
          "roomType",
          "availableCount",
          "minNetArea",
          "maxNetArea",
          "startingPrice",
          "currency",
          "minFloor",
          "maxFloor",
          "orientations",
          "floorPlanImageUrl"
        ]
      },
      "PublicPaymentPlanDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "downPaymentPercent": { "type": "string" },
          "termMonths": { "type": "number" },
          "deliveryPercent": { "type": "string" },
          "isRecommended": { "type": "boolean" },
          "monthlyPayment": { "type": "string", "nullable": true },
          "totalPrice": { "type": "string", "nullable": true },
          "cashDiscountPercent": { "type": "string", "nullable": true },
          "timelineNote": { "type": "string", "nullable": true }
        },
        "required": [
          "id",
          "name",
          "downPaymentPercent",
          "termMonths",
          "deliveryPercent",
          "isRecommended",
          "monthlyPayment",
          "totalPrice",
          "cashDiscountPercent",
          "timelineNote"
        ]
      },
      "PublicProjectMediaDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "type": "string", "enum": ["IMAGE"] },
          "url": { "type": "string" },
          "altText": { "type": "string", "nullable": true }
        },
        "required": ["id", "kind", "url", "altText"]
      },
      "PublicProjectAmenityDto": {
        "type": "object",
        "properties": { "code": { "type": "string" }, "label": { "type": "string" } },
        "required": ["code", "label"]
      },
      "PublicProjectPointOfInterestDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "category": { "type": "string" },
          "distanceMeters": { "type": "number" }
        },
        "required": ["id", "name", "category", "distanceMeters"]
      },
      "ProjectDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "developerOrganizationId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "developerName": { "type": "string" },
          "developerVerified": { "type": "boolean" },
          "status": { "type": "string", "enum": ["DRAFT", "IN_REVIEW", "PUBLISHED", "ARCHIVED"] },
          "constructionStatus": {
            "type": "string",
            "enum": ["PLANNED", "UNDER_CONSTRUCTION", "READY"]
          },
          "province": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "district": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "latitude": { "type": "string", "example": "40.990868" },
          "longitude": { "type": "string", "example": "29.027707" },
          "startingPrice": { "type": "string", "example": "7500000.0000" },
          "currency": { "type": "string", "example": "TRY" },
          "deliveryDate": { "type": "string", "format": "date", "nullable": true },
          "summary": { "type": "string", "nullable": true },
          "heroImageUrl": { "type": "string", "nullable": true },
          "stockUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "priceUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "version": { "type": "number" },
          "unitTypes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProjectUnitTypeDto" }
          },
          "paymentPlans": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicPaymentPlanDto" }
          },
          "media": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicProjectMediaDto" }
          },
          "amenities": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicProjectAmenityDto" }
          },
          "pointsOfInterest": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PublicProjectPointOfInterestDto" }
          },
          "availableUnitCount": { "type": "number" }
        },
        "required": [
          "id",
          "developerOrganizationId",
          "name",
          "slug",
          "developerName",
          "developerVerified",
          "status",
          "constructionStatus",
          "province",
          "district",
          "latitude",
          "longitude",
          "startingPrice",
          "currency",
          "deliveryDate",
          "summary",
          "heroImageUrl",
          "stockUpdatedAt",
          "priceUpdatedAt",
          "createdAt",
          "updatedAt",
          "version",
          "unitTypes",
          "paymentPlans",
          "media",
          "amenities",
          "pointsOfInterest",
          "availableUnitCount"
        ]
      },
...
      "DeveloperProjectDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "developerOrganizationId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "developerName": { "type": "string" },
          "developerVerified": { "type": "boolean" },
          "status": { "type": "string", "enum": ["DRAFT", "IN_REVIEW", "PUBLISHED", "ARCHIVED"] },
          "constructionStatus": {
            "type": "string",
            "enum": ["PLANNED", "UNDER_CONSTRUCTION", "READY"]
          },
          "province": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "district": { "$ref": "#/components/schemas/ProjectLocationReferenceDto" },
          "latitude": { "type": "string", "example": "40.990868" },
          "longitude": { "type": "string", "example": "29.027707" },
          "startingPrice": { "type": "string", "example": "7500000.0000" },
          "currency": { "type": "string", "example": "TRY" },
          "deliveryDate": { "type": "string", "format": "date", "nullable": true },
          "summary": { "type": "string", "nullable": true },
          "heroImageUrl": { "type": "string", "nullable": true },
          "stockUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "priceUpdatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "version": { "type": "number" },
          "unitCount": { "type": "number" },
          "availableUnitCount": { "type": "number" },
          "leadCount": { "type": "number" },
          "paymentPlanCount": { "type": "number" },
          "completenessPercent": { "type": "number", "minimum": 0, "maximum": 100 },
          "mediaCount": { "type": "number" }
        },
        "required": [
          "id",
          "developerOrganizationId",
          "name",
          "slug",
          "developerName",
          "developerVerified",
          "status",
          "constructionStatus",
          "province",
          "district",
          "latitude",
          "longitude",
          "startingPrice",
          "currency",
          "deliveryDate",
          "summary",
          "heroImageUrl",
          "stockUpdatedAt",
          "priceUpdatedAt",
          "createdAt",
          "updatedAt",
          "version",
          "unitCount",
          "availableUnitCount",
          "leadCount",
          "paymentPlanCount",
          "completenessPercent",
          "mediaCount"
        ]
      },
...
    }
  }
}